+++ src/components/ids-accordion/ids-accordion-header.ts
@@ -11,6 +11,7 @@ import styles from './ids-accordion-header.scss';
 import IdsText from '../ids-text/ids-text';
 
 import type IdsIcon from '../ids-icon/ids-icon';
+import type IdsAccordionPanel from './ids-accordion-panel';
 
 // Expander Types
 const EXPANDER_TYPES = ['caret', 'plus-minus'];
@@ -49,8 +50,10 @@ export default class IdsAccordionHeader extends Base {
 
   connectedCallback() {
     super.connectedCallback();
-    this.#refreshIconDisplay(this.icon);
+    this.refreshIconDisplay(this.icon);
+    this.#refreshExpanderIconType();
     this.refreshDepth();
+    this.refreshPlacement();
   }
 
   /**
@@ -120,6 +123,21 @@ export default class IdsAccordionHeader extends Base {
     }
 
     return false;
+  }
+
+  /**
+   * @returns {boolean} if this header component
+   */
+  get siblingsCanExpand() {
+    const panel = this.panel;
+    const parent = <IdsAccordionPanel>panel.parentElement;
+
+    let canExpand = false;
+    parent.childPanels?.forEach((childPanel: IdsAccordionPanel) => {
+      if (!canExpand && childPanel.isExpandable) canExpand = true;
+    });
+
+    return canExpand;
   }
 
   /**
@@ -197,8 +215,8 @@ export default class IdsAccordionHeader extends Base {
     const cl = this.container?.classList;
     const oldTypeClass = `expander-type-${oldType}`;
     const newTypeClass = `expander-type-${newType}`;
-    cl?.remove(oldTypeClass);
-    cl?.add(newTypeClass);
+    if (oldType) cl?.remove(oldTypeClass);
+    if (newType) cl?.add(newTypeClass);
   }
 
   /**
@@ -226,14 +244,14 @@ export default class IdsAccordionHeader extends Base {
         this.setAttribute('icon', `${val}`);
       }
 
-      this.#refreshIconDisplay(val);
+      this.refreshIconDisplay(val);
     }
   }
 
   /**
    * @param {string} val the icon definition to apply
    */
-  #refreshIconDisplay(val: string | any[] | null) {
+  refreshIconDisplay(val: string | any[] | null) {
     const iconDef = typeof val === 'string' && val.length ? val : '';
     const iconElem = this.container?.querySelector<IdsIcon>('.ids-accordion-display-icon');
 
@@ -245,7 +263,12 @@ export default class IdsAccordionHeader extends Base {
     }
 
     const hasParentIcon = this.parentHasIcon;
+    const siblingsCanExpand = this.siblingsCanExpand;
+    const expandable = this.panel.isExpandable;
+
     this.container?.classList[hasParentIcon ? 'add' : 'remove']('parent-has-icon');
+    this.container?.classList[expandable ? 'add' : 'remove']('is-expandable');
+    this.container?.classList[siblingsCanExpand ? 'add' : 'remove']('siblings-can-expand');
   }
 
   /**
@@ -350,6 +373,7 @@ export default class IdsAccordionHeader extends Base {
       iconType = this.expanded ? ICON_PLUS : ICON_MINUS;
     }
     icon.setAttribute('icon', iconType);
+    this.#refreshExpanderIconClass(null, this.expanderType);
   }
 
   /**
@@ -435,5 +459,16 @@ export default class IdsAccordionHeader extends Base {
 
     this.depth = depth;
     if (depth > 0) this.container.classList.add(`depth-${depth}`);
+  }
+
+  refreshPlacement() {
+    if (!this.container) return;
+
+    const panel = this.panel;
+    const isFirstChild = !panel.previousElementSibling || (panel.previousElementSibling && panel.previousElementSibling.tagName !== 'IDS-ACCORDION-PANEL');
+    const isLastChild = !panel.nextElementSibling;
+
+    this.container!.classList[isFirstChild ? 'add' : 'remove']('first-child');
+    this.container!.classList[isLastChild ? 'add' : 'remove']('last-child');
   }
 }