+++ src/components/ids-button/ids-button.ts
@@ -91,6 +91,7 @@ export default class IdsButton extends Base {
     this.#setInitialState();
     this.shouldUpdate = true;
     if (this.getAttribute('width')) this.width = this.getAttribute('width');
+    this.cssClass = this.cssClass;
   }
 
   #setInitialState() {
@@ -104,6 +105,9 @@ export default class IdsButton extends Base {
     if (this.hasAttribute(attributes.CONTENT_ALIGN)) this.setContentAlignClass(this.getAttribute(attributes.CONTENT_ALIGN));
     this.setIconAlignment();
     this.refreshProtoClasses();
+
+    // set initial direction
+    this.localeAPI.updateDirectionAttribute(this, this.localeAPI.language.name);
   }
 
   /**