+++ tests-to-convert/ids-input/ids-input-e2e-test.ts
@@ -27,7 +27,7 @@ describe('Ids Input e2e Tests', () => {
   it('should pass Axe accessibility tests', async () => {
     await page.setBypassCSP(true);
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
-    const results = await new AxePuppeteer(page).analyze();
+    const results = await new AxePuppeteer(page).disableRules(['color-contrast']).analyze();
     expect(results.violations.length).toBe(0);
   });
 });