+++ webpack.config.dev.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const sass = require('sass');
+const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const demoEntry = require('./scripts/webpack-dev-entry');
@@ -173,6 +174,10 @@ module.exports = {
           }
         },
       ]
-    }),
+    })
   ].concat(htmlExamples)
 };
+
+if (isProduction) {
+  module.exports.plugins.push(new BundleAnalyzerPlugin());
+}